requirements.txt
discord.py
python-dotenv
aiohttp
pyngrok
mysql-connector-python
asyncio